# optimize decoding of large-chunk 