--- a/benchmark/jsonresultset/.gitignore
+++ b/benchmark/jsonresultset/.gitignore
@@ -1,0 +1,3 @@
+*.test
+*.out
+prof_largeresults