### Description
From the primary commit message.

>    The primary goal is to reduce to in-flight memory
    consumption to be as low as possible. This is done
    by making assumptions about the chunk structure
    and using a streaming decoder.
    
>    The secondary goal is to optimize the CPU cycles
    needed to decode the chunk. There's still more
    to do on this front, but the design in this patch
    should be conducive to future optimizations.

The impetus for this change was primarily to reduce memory pressure when downloading large `json` results. However this optimization should reduce pressure for all large results. For example, I see the following improvement for the existing large-result benchmark.

```
-- before change
BenchmarkLargeResultSet-4 1    68703150935 ns/op   6451121456 B/op    270103889 allocs/op
-- after change
BenchmarkLargeResultSet-4 1    57140518986 ns/op   3198717528 B/op    220123790 allocs/op
```

For the new `json` case, I see the following improvement. It's worth noting that I ensured my test runs used the same cached result; future comparisons may see different numbers if the cached rowset (in `S3`) is different.

```
-- before change
BenchmarkJsonResultSet-4 1  114852005941 ns/op  2964585432 B/op  149820 allocs/op
-- after change
BenchmarkJsonResultSet-4 1  68804250571 ns/op   1056399984 B/op  288150 allocs/op
```

It's worth mentioning that before hand-writing a `json` decoder, I tried a few different libraries. None of them performed as well as I would have liked, and one of the streaming libraries actually performed worse (with regards to memory pressure) than `encoding/json`. 

### Checklist
- [ ] Code compiles correctly
- [ ] Run ``make fmt`` to fix inconsistent formats
- [ ] Run ``make lint`` to get lint errors and fix all of them
- [ ] Created tests which fail without the change (if possible)
- [ ] All tests passing
- [ ] Extended the README / documentation, if necessary