@smtakeda 2018-07-29
Thank for PR. Will look into this after coming back from off.