@smtakeda 2018-08-12
Other than those two test cases, looks good to me.