@smtakeda 2018-08-12
In my verification tests, the memory footprint was reduced from 2G to 600MB, while no much difference in CPU usage.