@smtakeda 2018-08-12
hmm, for some reason, the test doesn't pass:
https://travis-ci.org/snowflakedb/gosnowflake/builds/414998093