@smtakeda 2018-08-12
Ran TestLargeSetResult and got the log:
https://travis-ci.org/snowflakedb/gosnowflake/jobs/415002366

The cause is
```
"Client.Timeout exceeded while reading body"
```
Adjusting the buffer size, timeout, downloader goroutine...