@mhseiden 2018-08-13
Huh, that's funky. Even when running the large result set tests on a bad network, I wasn't running into timeout issues 🤔