@smtakeda 2018-08-14
It worked ok on my dev VM (3 virtual cpu) and MacOS (4 cores), but it is not on Travis (2 cores). I speculate the decoder walks through the response body incrementally, so it takes times to read whole data and hit timeout.

Changing the max number of chunk downloader worker from 10 to 2 and increasing the client timeout to 900 worked, but the performance significantly degraded.

I'll look into the parameters further, but in either case, if we want to have the custom decoder, I will have a feature flag to switch between the native JSON and the custom decoders for the applications to choose. At this time, I feel it is a bit risky to completely switch to the new decoder.