@mhseiden 2018-08-14
Yea, adding a feature flag definitely makes sense. I'm happy to add that in, unless you're already working on it?