@smtakeda 2018-08-16
Wrapping up in https://github.com/snowflakedb/gosnowflake/pull/184